refactor: move thread synthesis into AgentContext.fromPacket
Move the thread/channel override logic out of the bot-layer helper
and into AgentContext.fromPacket as an options argument. Context
construction now lives entirely in the AI context module; the bot
layer only owns the Discord-specific recent-messages fetch.

Callers compose the two explicitly:
  const recentMessages = await fetchRecentMessages(...);
  const ctx = AgentContext.fromPacket(packet, {
    threadOverride,
    recentMessages,
  }).toJSON();

Renames src/bot/chat-context.ts to src/bot/recent-messages.ts to
reflect its narrowed responsibility.
