+++ src/lib/ai/context.test.ts
@@ -66,6 +66,33 @@ describe("AgentContext.fromPacket", () => {
   it("sets a formatted date string", () => {
     expect(AgentContext.fromPacket(messagePacket("hello")).date).toMatch(/\w+, \w+ \d+, \d{4}/);
   });
+
+  it("synthesizes thread info from threadOverride", () => {
+    const ctx = AgentContext.fromPacket(messagePacket("hello"), {
+      threadOverride: { id: "thread-99", name: "my-thread" },
+    });
+    expect(ctx.channel).toEqual({ id: "thread-99", name: "my-thread" });
+    expect(ctx.thread).toEqual({
+      id: "thread-99",
+      name: "my-thread",
+      parentChannel: { id: "ch-1", name: "general" },
+    });
+  });
+
+  it("threadOverride takes priority over packet thread info", () => {
+    const ctx = AgentContext.fromPacket(
+      messagePacket("hello", { thread: { parentId: "p1", parentName: "parent" } }),
+      { threadOverride: { id: "thread-99", name: "new" } },
+    );
+    expect(ctx.thread?.id).toBe("thread-99");
+    expect(ctx.channel.id).toBe("thread-99");
+  });
+
+  it("attaches recentMessages when provided", () => {
+    const messages = [{ author: "bob", content: "hi", timestamp: "1:00 PM" }];
+    const ctx = AgentContext.fromPacket(messagePacket("hello"), { recentMessages: messages });
+    expect(ctx.recentMessages).toEqual(messages);
+  });
 });
 
 describe("AgentContext serialization", () => {
