+++ src/server/routes/handlers.ts
@@ -4,11 +4,12 @@ import { resumeHook } from "workflow/api";
 import type { EventHandler } from "@/bot/events/types";
 import type { ChatHookEvent } from "@/workflows/chat";
 
-import { buildTurnContext } from "@/bot/chat-context";
 import * as userEvents from "@/bot/handlers/events";
 import { handleMention } from "@/bot/handlers/events";
 import { isBotMention } from "@/bot/mention";
+import { fetchRecentMessages } from "@/bot/recent-messages";
 import { EventRouter } from "@/bot/router";
+import { AgentContext } from "@/lib/ai/context";
 
 export const router = new EventRouter();
 
@@ -28,7 +29,8 @@ router.onMessage(async (packet, ctx) => {
   log.info("handlers", `Forwarding message to workflow ${existing.workflowRunId}`);
 
   try {
-    const turnContext = await buildTurnContext(ctx.discord, packet);
+    const recentMessages = await fetchRecentMessages(ctx.discord, channelId, packet.data.id);
+    const turnContext = AgentContext.fromPacket(packet, { recentMessages }).toJSON();
     const event: ChatHookEvent = {
       type: "message",
       content: packet.data.content,
